fix(stream): use setTimeout over setImmediate
setTimeout is platform agnostic and webpack 5 will no longer polyfill setImmediate for us.

This commit also refactors some of the tests to wait for specific signals/events rather than waiting a specific number of cycles of the event loop. This should be more robust and less prone to breaking due to changes in Node.js internals.